requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langsmith>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.38.0
